+++ src/Changes/Manual/BalrondContainers.cs
@@ -45,6 +45,26 @@ public class BalrondContainers : ManualChangesBase
 
     var potionShelf = this.PieceManager["piece_chest_shelf_bal"];
     this.SetContainerRestrictions(potionShelf.Prefab, potionShelf.Prefab.transform.Find("potionslot-1"));
+
+    // Fix comfort
+    string[] removeComfort = [
+      "piece_chest_fabricshelf_bal",
+      "piece_chest_shelf_bal",
+      "piece_chest_weaponrack",
+    ];
+    foreach (var piece in this.PieceManager[removeComfort]) piece.Comfort.Value = 0;
+
+    string[] basicTablesComfort = [
+      "$piece_chest_foodtable_fine",
+      "$piece_chest_foodtable",
+    ];
+    foreach (var piece in this.PieceManager[basicTablesComfort]) (piece.Comfort.Value, piece.Comfort.Group) = (1, Piece.ComfortGroup.Table);
+
+    string[] fancyTablesComfort = [
+      "$piece_chest_foodtable_round",
+      "$piece_chest_foodtable_oak",
+    ];
+    foreach (var piece in this.PieceManager[fancyTablesComfort]) (piece.Comfort.Value, piece.Comfort.Group) = (2, Piece.ComfortGroup.Table);
   }
 
   private void SetContainerRestrictions(GameObject prefab, Transform slot)