+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix Balrond Containers comfort values.
+
 ## [0.6.36] - 2024-07-18
 
 ### Fixed